wip commit

+++ db/migrate/20200714122642_remove_user_devise.rb
@@ -11,11 +11,11 @@ class RemoveUserDevise < ActiveRecord::Migration[6.0]
     remove_column :users, :reset_password_sent_at
     remove_column :users, :allow_password_change
 
-    remove_column :users, :sign_in_count
-    remove_column :users, :current_sign_in_at
-    remove_column :users, :last_sign_in_at
-    remove_column :users, :current_sign_in_ip
-    remove_column :users, :last_sign_in_ip
+    remove_column :users, :sign_in_count if column_exists? :users, :sign_in_count
+    remove_column :users, :current_sign_in_at if column_exists? :users, :current_sign_in_at
+    remove_column :users, :last_sign_in_at if column_exists? :users, :last_sign_in_at
+    remove_column :users, :current_sign_in_ip if column_exists? :users, :current_sign_in_ip
+    remove_column :users, :last_sign_in_ip if column_exists? :users, :last_sign_in_ip
 
     # DeviseTokenAuth attrbiutes
     remove_index :users, %i[uid provider]
@@ -27,6 +27,6 @@ class RemoveUserDevise < ActiveRecord::Migration[6.0]
     # User attributes
     remove_column :users, :first_name
     remove_column :users, :last_name
-    remove_column :users, :username
+    remove_column :users, :username if column_exists? :users, :username
   end
 end
